Removed dependency on libevent internals.
We now track the link between an event base and vevent base using a hash
table instead of storing the pointer inside the event base itself. This
means we do not need to know about the internals of the event base
struct. There is probably a better way to do it, but this works for now.
Closes #2.