Fix echo-eth0 example.
The client in the built-in echo example now receives entire echo
replies consistent with what it sent. The echo server correctly
buffers data that it reads but can not immediately send due to CPU
or bandwidth constraints. Fixed in branch release-1.0.0 (which will
eventually make its way to master). Fixes #4.