Fix for spontaneous crashes
Don't add to the reference count when retransmitting packets, because
the event tracker doesn't actually hold the packet (it copies data
that it needs out of the packet). Added some code so shadow DIEs if
the reference count is "out of bounds" (<0 or >100). The bug happened
because of an overflow of the count which caused the packet to be
destroyed, then accessed later.