Sane sizes for default network topology.
Reduced number of relays, clients, and servers so the default topology
runs within 3 GB of Ram. Downloads are completing with scallion.
Also, changed default location to ~/.local so it doesn't require root,
and is easier to sync up the build with the run. I.e. you dont have to
use -p when running after installing with auto.