Generic plug-in callbacks in plug-in library context
It is useful and necessary to allow vevent to callback into the plugin
code. However, to ensure the plug-in context is set up properly this
should go through plugin. So we add a generic callback to enable this.
vevent was modified to use the generic callback instead of the direct
calls it made before.

refs #28