Fixes for preload and intercept libraries.
Added new functionality that allow us to determine if a function call
came from the plug-in or from shadow. This is useful so that glib can
access files, etc, correctly even if we normally intercept those
functions. To do this we need to track our context whenever we do a
context switch. There are new macros to do this, and to decide wheather
or not to redirect a system/library call to shadow.