Major bugfix for stalled sockets when blocked on CPU.
When read/write calls block on CPU, we did not track this with vepoll.
This means when the CPU delay was absorbed, we failed to notify the
plug-in that the socket is available. This has been fixed.

As a side effect, we also fixed the CPU delay check until after we
determine if the socket is a pipe. This means pipes essentially have no
cpu delay.

Also updated the pingpong.xml file.