Make plug-ins open with global symbols
only global dlopens are searchable with dlsym. we cant use
G_MODULE_BIND_LOCAL if we want to be able to lookup functions using
dlsym in the plugin itself. if G_MODULE_BIND_LOCAL functionality is
desired, then we must require plugins to separate their intercepted
functions to a SHARED library, and link the plugin to that.