Defer socket descriptor loop-around problem.
We store virtual socket descriptors as uint16 for some reason. When
running a long experiment, or on large topologies, some nodes descriptor
counter may loop around, causing problems with pipecloud simulator
functionality. For now we lower the starting point of the counter to
delay the problem.

A better solution is to switch to a 32-bit virtual socket descriptor, so
I'm leaving this issue open until that is resolved.

refs #22