Toward Tor working in new shadow
This adds the refill interval functionality from tor-0.2.3.5-alpha that
currently breaks any version of Tor before that one. Need to eventually
handle this properly.

Also added a quick script to install dependencies that used to be
installed as part of shadow, but now will be as part of scallion.

refs #10