New CPU delay model
Implemented new CPU delay model, where processing delays from the
application are measured when the plugin transfers control to the app
and when control returns to shadow. This delay is incorporated into the
CPU for that node based on its CPU frequency. Added a configuration
option for the delay threshold below which delay is ignored.

Still need to test if this has any effect on busy nodes, and incorporate
it into the XML structure.

refs #34