Track node network activity and plug-in cpu utilization
A new tracker class tracks node statistics. Currently, we are tracking
CPU processing times for each node application (time spent executing
inside the plug-in) and bytes in and out of each network interface. 

This also includes some code to track memory used by the plug-in, but
the malloc family of functions that we normally would intercept for this
feature are commented out due to initialization errors when preloading.

Takes care of 1, 2, and most of 3 in issue #40