Change default tracker heartbeat to "message" level
The stats that are collected by our internal node trackers are extremely
useful in analyzing the state of the network and what happened in an
experiment. This information should be logged at message level, and
should always be printed by default.

Also, make it easier to identify the heartbeat message as shadow's.