Keep socket readable when EOF signal is needed
This is an important bugfix for some the slow bootstrapping problems
that have been plaguing our Tor experiments. A lot of communication with
the directory servers (and others?) depends on reading EOF from the
socket before taking the next step (like during consensus downloads).

There was a corner case where the socket had received a remote close
signal (FIN), and afterwords the user read all of the remaining data
out of the socket. In this case we were marking the socket as unreadable
because there was no data, when in fact it needed to be marked as
readable so the user could read the EOF and close the socket.

I'm not closing the issue yet in case I find other issues in the next
few days.

refs #62