Towards new threading model
Added new concurrent utilities and abstracted the node event queue.
The "tree of queues" design is implemented, but not currently used in
this version because I'm skeptical that it actually improves anything
over a simple async pqueue. I'm adding it so a version of the code
exists, although some fixing is needed if that design is wanted in the
future.

refs #88