Implemented Tor controller that can interact with scallion nodes -- Issue #98
Changes:
--------
* Updated scallion code to intercept function which adds callbacks to
log messages.  This enables log messages to be forwarded to the Tor
controller.
* Added some parameters that get automatically set when starting Tor.
These include setting the control port to 9051, having Tor listen on the
hostnames address so other nodes can connect, and setting the password
for the Tor controller to "password".
* Implemented a base Tor controller plugin that is able to send commands
to the scallion plugn, as well as parse some of the replies sent back.
* Impemented a module that can create custom circuits for nodes.

Notes:
------
In order to use, just one node using the plugin
"libshadow-plugin-tor-control.so" needs to be started.  The only
argument it takes in is a filename to a tor control hosts file.  Each
line of the hosts file contains the node to connect to, the module to
load, and any arguments that module takes.  For example, to use the
circuit build module, here is a sample line that should appear in the
file:

webclient1:9051 circuitBuild nonexit1 nonexit2 exit1 -1

For the circuit build module, circuits can be of length 1 to 8, and the
list must end in a "-1".  Note that if the last hop is not normally and
exit node, you must make sure the node's torrc file does not have a
"reject *:*" exit policy, other wise the circuit will not work.  In
addition, if you want to create one hop nodes, you must make sure every
torrc file has these config settings:

ExcludeSingleHopRelays 0
AllowSingleHopCircuits 1
AllowSingleHopExits 1

Otherwise nodes will be unable to bootstrap and be used.