Changed the latency model and topology dataset
The dataset was gathered from iPlane (http://iplane.cs.washington.edu/)
using the planetlab to all destination traceroutes.  Collected 40 days
of data and computed the following: minimum, 1st quintile, mean, median,
3rd qunitile, maximum and standard deviation.  The median and standard
deviation are stored in the "latency" and "jitter" fields in the
topology file, while the rest were added for the new model.  If there
was incomplete data, the data from the old topology file was used (which
also used the old sampling method).

For computing the latency now, the percentile determines which range to
use:

* 0 <= p <= 0.25: sample from [min, q1]
* 25 < p <= 0.50: sample from (q1, median]
* 50 < p <= 0.75: sample from (median, q3]
* 75 < p <= 0.100: sample from (q3, max]

where the sampling is done uniformly.  This results in a tigher latency
distribution, where the existence of outliers towards max won't get
chosen nearly as often.

refs #27