Stop preloading EVP_Cipher by default
Preloading EVP_Cipher() is causing MAC errors on some machines with a
specific cipher mode that requires padding. Our memmove implementation
doesn't handle padding. There is a new option in the setup.py script for
enabling the preload for machines in which the padding bug is not
triggered.

Fixes #97

cherry-pick

Conflicts:
	CMakeLists.txt