Added default value of 1 to quantity for all nodes
This way you can explitly set quantity to be 0 if you want to have scripts modify hosts file to change what nodes are going to run.  This fixes the bugs seen in the file and echo plugin, where quantity was not explicitly set and thus the nodes were not being created.