Assign nodes to worker threads, remove thread pool
The old model allowed nodes to run in different threads for any given event. This causes inconsistent address space issues. For example, if a node stored an address of a function while running in the context of one thread's plugin, and later tried to call the function at that address while running in a different thread's plugin, it would call into the first threads address space and interfere with the other node's state.

By assigning virtual nodes to specific threads, they will always execute in the same plug-in context and should avoid these issues.

refs #56