New torcontrol relay ping module
This module extends a new circuit to the relay specified in the XML
program args, and measure the time until the circuit has been extended.
The circuit is then torn down. This can be used to measure the
approximate RTT through the relay.

NOTE: the torcontrol application must be started with enough delay after
the Tor client is started to give the client enough time to download the
relay descriptors.