better handling of tor logging
instead of intercepting Tor's logv(), use the add_callback_log()
to register a log callback that we can then use to direct log
messages through Shadow's logging facilities.

this happens automatically, so we no longer have to tell Tor to
log to stdout. tor's log messages are currently filtered based
on shadow's configured log level, but scallion can take a log
level input if we need that to change.

refs #128