fix format string error
The old format string for this log message only worked on machines
which can handle guint64 with a '%u' format specifer. This format
is incorrect and not generally compatible. The fix is to correctly
use glib's defined G_GUINT64_FORMAT specifer, which will always
define the correct format for the machine.

refs #140