set shadow context properly, fix malloc pointer tracking
shadow context was not set properly for all system_* functions,
meaning there were cases where malloc was getting called and
intercepted and we thought it was the plug-in calling it instead
of shadow. this has been updated to function correctly.

malloced pointers are now removed from the hash table on free

refs #144