update packetloss model
since shadow uses:
(1-srcloss)(1-linkloss)(1-dstloss)
then we need to set the srcloss and dstloss as sqrt of the values
given by netindex. this way, when src and dst are in the same
vertex and linkloss is 0, we end up with the original value
that was given by netindex.

also use median of all loss values instead of mean in order
to resist outliers.

refs #27