new retransmit queue, timers, and rtt estimates
-added packet status debug strings for easier tracking
-skip packet debug status if debug log level is filtered
-use priority queues for storing ordered packets
-significantly overhauled retransmission model to use timeouts
-updated packet ref/unrefs to follow the new retransmit model

This worked on the filetransfer with 90 percent packet loss,
and on Tor with 8 percent packet loss, though the time to download
files went up due to the packet loss.

The RTT estimate is still wrong. We need to avoid using invalid
timestamps when packet loss occurs, and reset the RTT state
variables after several backoffs occurred (both allowed under
rfc 6298). This is coming soon.

refs #185