integrating new cubic algorithm, and reno fixes
the cubic algorithm is the default we should use going forward.
reno contains some fixes to the original aimd algorithm that
shadow used.

currently, the cubic algorithm does not seem to work perperly
with the new retransmission design.

the congestion module refacting and the cubic algorithm are
written by John Geddes (@jdgeddes).

refs #145