WIP support for FDs less than 1024
This is an initial attempt for Shadow to return file descriptors
less than 1024. We have to deal with a plug-in that may open
files since the OS will also return low numbered FDs. Our
approach is to open those files through Shadow so that Shadow
can create a mapping between OS and shadow FD to prevent
duplicates.

Current progress is that preloading open causes problems for
Shadow during the Topology initialization. When open is not
preloaded, or directed through the OS only, the --file
test completes successfully.

refs #214