massive refactoring of tgen
server and event io functionalities have been moved to their own
files. notification callback functions were added so parts of
the code can subscribe to events occuring in other parts.

as a result of the notify funcs, this commit introduces known
problems with the ref/unref sync and will leak memory because
objects are not freed properly. but the code builds and seems
to run correctly otherwise.

refs #124