finished minimal socks proxy support, update peer
lookups for complete peer information will now only be performed
right before the direct connection to the peer occurs. this way,
if a proxy is being used to connect to a particular peer, that
peer's address will be retreived by the proxy and will not be
leaked to the network by us.

this commit should allow .onion hostnames to be used with tgen
when a proxy is specified, as we will simply send a socks connect
request to the proxy with whatever hostname was given in the
tgen graphml config file.

refs #124