refactor event system into a callback tasks and events
tasks can be created be giving a callback function and callback
params, and then tasks can be attached to an event that will be
executed at a specific time. this is cleaner and simpler than
creating a new event file/module for each event type.

this should be a refactor only; though most of the old event
files have been removed since they are no longer necessary.

refs #281