major refactor of cli options and xml config, replaced event/action system
parsing the configuration options no longer results in actions getting
created on the scheduler. instead, the master is responsible for
taking the xml config and acting on it to distribute the hosts, plugins,
etc to the appropriate slaves. it uses the cli options as defaults which
get overridden in the case that identical xml configs are supplied.
this has resulted in a simplification of the startup process.

virtual host and process bootstrapping has been separated from the
creation of instances of these modules.

also, topology and dns modules are global and now master owns them.
this will require updating when moving to multiple slaves.

closes #281