initial simple phold type benchmark
the test is initialized with a message load L. during the init
phase, each node will pick a node at random and send 1 byte
to it, and repeat this L times. during the execution phase,
each node will pick a random node and send it 1 byte for every
message it receives.

this was implemented as a shadow plug-in, but can be run using
the built-in '--test' command line option.

closes #280