interpose pthread library, add pthmt as dependency
we added the pthread interface functions to the preload interposer
file, with some pth-specific handling code for those functions
(still commented). we also set up shadow to link to pthmt correctly
and start using its threads to manage plugin execution contexts
for each virtual host. finally, we changed the way the tgen plugin
is built, so we can now have a single dynamic executable that can
both be run outside of shadow and run inside of shadow; this will
enable consistent funcionality, and eventually shadow will just
call the main function of the plugin.

refs #236