fully utilize epoll and timer fds
rpth now uses a epollfd for each thread, and a global epollfd to
watch each of the thread epollfds and associated events. timerfd
is used to faciliated internal rpth timer functions for sleep and
other timeouts. finally, we added a new mode to the rpth global
context to specify that rpth should never block, and added a new
rpth scheduler loop that is used when this mode is active.