prevent tor startup issues when running shadow with multiple threads
there is a race condition in the tor init process when running
shadow with multiple worker threads. the recent changes to the tor
plugin to handle openssl and libevent global locking do not seem
to do enough to prevent the race. we are missing some part of
the init process that needs ot be locked to be thread safe.
this is a temporary hack to lock the entire tor init process from
within shadow, which seems to remedy the problem.