don't rely on igraph to be thread-safe yet
preliminary tests, starting 2000 tgen nodes at the same time,
caused the nodes to stall during startup without any indication of
error. this indicates that igraph is not properly thread safe. so
lets not yet rely on igraph's thread safety.

the code that is required to utilize a thread-safe igraph is in,
although never excercized as of this commit. still, i'm closing
the issue, and hopefully we can simply reverse this commit at a
later point.

closes #259