avoid glib recursive log errors
when plugins log messages using glib, and shadow runs out of
free file descriptors, it would be nice to log a warning message.
if we do this with the usual warning() shadow log method, glib
freaks out because it thinks we have a recursive log.
this hack gets around that issue, but is not the ideal way to
handle the issue.