Add new option to build shadow plugin as exe
Shadow can load and run plugins that are built as position
independent executables as well share libraries. We added a new
macro to build the PIE version. The PIE is more robust since it
can be executed outside of Shadow too, but the internal symbols
inside the PIE can not be interposed by the LD_PRELOAD trick.
Therefore, we leave the shared library option for plugins that
find internal symbopl interposition useful (the Tor plugin).