Add optional plug-in hook functions
Shadow will call the new hook functions if they are defined in the
plug-in library. This allows the plug-in to any necessary
adjustments to running more smoothly in Shadow.

The symbols Shadow searches for and will call are:
__shadow_plugin_load__
__shadow_plugin_unload__
__shadow_plugin_enter__
__shadow_plugin_exit__

Shadow will pass a unique void* pointer in these calls. The GLib
GModule* of the plug-in library is currently passed.

refs shadow/shadow-plugin-tor#22