Make the version requirement checks more closely follow glibc's behavior. The changes here include: * Removing the ability to find version requirements in the verdef array. The verdef array is only for providing versions of symbol definitions, not symbol version requirements. When looking up a symbol, we should never require a version from the verdef array.
* Removing the ability to find version definitions in the verneed array.
Similarly, the verneed array is only for giving symbol version requirements. The code finding symbol version definitions in the verneed array was "buggy" in the sense that it shouldn't be happening at all. This, in combination with the next point, was the source of some of the bugs we have been working around.

* Correctly determine if there is no version definition.
Prior to this change, the only way to trigger the code path where the potential match is not considered to be versioned is if there was no versym table at all (which would only happen if there were no symbol version definitions *and* no symbol version requirements). With this change, the symbol is considered a match if there is no verdef array or there is no versym array. We also add a case to check if there is a verdef array but this particular symbol has no version --- a case I have yet to see, but is accepted in glibc.

* Remove is_preloaded workaround
We no longer have a need for this field.