Merge plugin into process, correctly set context on load
When plugins are loaded with dlmopen, their init functions are
called. Those init functions may initialize some memory, and so
we need to call those functions while in the plugin context to
make sure we direct any functions intercepted during plugin
initialization to the correct versions.

Prior to this, plugins that use glib would fail, because glib was
setting some thread local keys during the constructor which we
previously executed in the Shadow context. Later when running the
plugin, the glib key would no longer exist when we looked for it
inside pth.