WIP adding support for loading preload libs at runtime
this required loading and validating the configuration file
earlier in the startup process than usual, which required
starting the logging system sooner. this is still not in its
final state and is definitely rough around the edges, but
should give an idea of what would be needed to load the
preload libs.