reduce memory leak in timer
When a timer is set to expire far in the future, we used to schedule
an event and forget about it until it expired. This commit changes
the functionality so that we check on the event every second to see
if the user closed the timerfd or disarmed it in the meantime. If
that happened, we can unref the timer ASAP.

This only reduces the total number of hanging timers to those
created in the last second. There is still a slight leak here,
because the timers that were created in the last second of the
sim will not be freed correctly when the sim ends. This is due
to a missing feature of the task object used internally to store
tasks in the scheduler - it has no concept of how to free the
data that it stores, and so anything held in task events will be
lost. I've opened #330 to track and deal with that issue.