reset retransmission timeout to defined init value
The code used to reset the RTO to 1ms. The RFC 6298 specifies to
reset it to the initial value of 1s. I *think* this was a coding
mistake: 1 millisecond was used instead of 1 second.

Setting to 1 ms as was done prior to this commit was causing an
effective min of 200ms when this was clamped. In any case, we are
now resetting to the initial value.