rename node and application elements
Changed the name of the 'node' element to 'host'.
Changed the name of the 'application' element to 'process'.
Also, refactored the 'time' attribute of the 'kill' element to be
specifiable as the 'stoptime' attribute of the 'shadow' element.

The old element names are still supported for the time being, but
will become deprecated at some unspecified point in the future.