stop using inferrior glib lists in config parser
The GLib list structure iterates the list on append and on
length checks. This was causing the config load process to take
an hour or more when loading a file of 100M with hundreds of
thousands of nodes.

Now this is refactored to use the GLib queue, which is double
ended and far more intelligent.