correct the calculation of required TLS size
When booting up, shadow now correctly counts the static tls size
needed for each namespace by loading the preload library for
a virtual process in the same namespace as the process plugin
library. This should reduce the amount of static tls required
because we won't double count libc or any other libraries that
are dependencies of both the plugin and the preload libraries.