work around pthread_join issue causing deadlocks
pthread_join may cause deadlocks in the new elf loader. This may
happen when a segfault occurs inside of the loader and then Shadow
tries to get a backtrace which enters the loader again while it is
holding a lock. We work around the issue by just letting the thread
end itself, after communicating that it has flushed the log back
to the main thread.