work around pthread_join for main shadow workers
Don't call pthread_join in the Shadow worker threads to prevent
deadlocks that may happen in the loader. The approach used here
to work around the issue is to use a count-down latch so that the
thread can notify the main thread that it has finished all
internal processing. This is the same general approach used for
logger helper thread.