stop plugin segfaults and aborts from killing shadow
Installed a signal handler for each plugin namespace for deadly
signals SIGSEGV, SIGFPE, SIGILL, and SIGABRT so that if those
signals are tiggerred in the plugin, Shadow will kill the pth
thread instead of allowing the entire Shadow process to be killed.

Started intercepting sigaction to prevent the plugin from setting
custom signal handlers for those deadly signals.

Started intercepting abort() to also kill only the pth thread.

Note: _process_unloadPlugin is commented out in two places and
tagged with "XXX TODO" to indicate that we need to add that back
in once elf-loader supports unloading.