start tracking execution time of each host
Track the amount of time a worker thread spends executing each
host. This total elapsed execution time for a host is available
in a new function call host_getElapsedExecutionTime().

This commit also fixes a bug where hosts never actually get
freed correctly because items internal to the freeing process
hold references to the host itself. This means the ref count
could normally never get to 0, and the host would not be freed.