quick stab at adding thread wait timers
Currently, the thread wait timers only time the time that elapses
while each thread is waiting for the round-end barrier, i.e.,
waiting for the other threads to finish executing their events
in a round so that everyone can proceed to the next round.
Ideally, we minimize the time waiting for this barrier, and
so we would distribute load among threads in a way that minimizes
the wait time for this barrier.