Fix epoll ONESHOT duplicate event reporting bug
If using ONESHOT, the events are reported multiple times even
though they should only be reported once.

Then if the event was collected and epoll_ctl is called with
EPOLL_CTL_MOD, the event would not be reported again when it
should be.

This commit adds test cases for both scenarios as well as fixes
to the epoll subsystem.

Thanks for reporting and fixing, @RWails!

fixes #349