Fix epoll edge-trigger event reporting
In epoll edge-triggered mode, Shadow would not report events after
an epoll_ctl MOD operation because the underlying descriptor did
not change. However, Linux treats a MOD operation itself as a
change in descriptor status and reports an event following a MOD.

This commit adds a unit test to Shadow's epoll test to exercise
this behavior and verify that Shadow is inconsitent from Linux,
and also adds a fix to the epoll subsystem to correctly report
events in edge-trigger mode following a MOD operation.

Thanks for reporting @bobbyvinon!

Fixes #343