Check topology completeness by counting edges at all vertices. (#359)
Testing this on a very large complete topology and the same topology after removing one edge, it no longer hangs. It also returns the correct answer. It just takes the usual ~15 minutes to load the darn thing. Verification of completeness is on the order of 30 seconds.

The idea is to count edges at every vertex. I think this works ... as long as you never have more than one edge between a pair of vertices.

closes #358