Intercept syscall, and handle clock_gettime syscall
This commit adds syscall() to the list of functions that Shadow
preloads. The syscall() variant can be used for many (most?) libc
functions, and so anyone using syscall() instead of the normal
libc function will likely run into issues.

We added support for clock_gettime, and a new test case that shows
that a c++ time function does indeed use syscall(). There might be
others.

The full list of syscall functions are listed in
src/main/host/shd-process.c in process_emu_syscall. We print a
critical message if a should-be-supported function is called with
syscall, and an info message otherwise.

Fixes #365