Add determinism test and fix fread on random files
Added a determinism test that opens/reads and fopens/freads from
random devices /dev/random and /dev/urandom. The test outputs
the random bytes that were read over two runs and ensure that
the output is the same.

We also fixed fread so that it correctly handles reading from
these random devices.

Fixes #368