better validation of graph attributes, shortest path for self-edges
This commit adds some new code to validate the types and existence
of graph, vertex, and edge attributes. It checks that the required
attributes exist in the graph. If validation fails, Shadow can
not run a simulation.

It also contains a significant refactoring of the shortest path
calculation code. Previously, a path from a vertex back to itself
would always use 1 millisecond of latency, even if the graph
itself did not have an edge to itself. In this case, we now find
the shortest path to any *other* vertex, and use that path twice
in order to get back to the source. Source-to-source edges are
preferred if they exist in the graph.