Stop accidentally double-ref-counting packets
There was a corner case where a retransmitted packet would arrive
twice, and attempted to be queued in the TCP unordered input
queue so that the app could read it. We counted the reference
twice, even though the queue only kept a single pointer to the
packet. This commit adds some better checks for membership in the
queue before counting another referernce.

Also, there are related memory leaks with TCP objects that are
causing retransmission queue packets to not be freed. This adds
a call to clear_retransmit when the socket closes. This does not
fix the fact the the TCP sockets themselves are not freed properly.

Lost packets have gone down dramatically, but there still seems
to be some packets that are not being freed.

refs #394