Fix refcounts in new tmodel schedule transfer type
For schedule transfer types, we were mis-using timer events and
tracking reference counts incorrectly. This caused transfer and
transport objects to not get freed and the corresponding sockets
to not get closed properly.

Schedule timers are also correctly canceled in the case where we
timeout or stallout before the transfer completes.