Fix schedule pause bug, more efficient schedule writing
This fixes a bug where the transfer schedules were only pausing
according to the delay of the last packet sent "at the same time"
(according the the packet delay threshold) rather than pausing an
amount of time equal to the cumulative delay of all packets that
were sent "at the same time".

The schedule writing code was also rewritten for efficiency.