Event compare uses event IDs instead of push sequences
When comparing events, when the source and destination host are
determined to be equal, we then sort the event based on the order
in which the event was created on the src host (i.e., the eventID).
This is done rather than sorting on the sequence order that
the worker thread pushed the event, which could potentially be
different across different experiments if the host was stolen to
a different thread.

In other words, making the comparison host-specific rather than
thread-specific means we sort independent of which threads the
hosts events are executed on.

refs #666