Make deep copy of packets when transferring across hosts
This fixes a bug that was occurring in the case that packets were
being retransmitted even though the receiver was also in the
process of receiving them. In this case, the sender was updating
the packet header info before sending out the retransmitted
packet, and the receiver was then referencing that updated info
from the original packet (i.e., before receiving the
retransmitted packet). This happens because packets are shared
among sender and receiver to save memory.

The fix introduced in this commit is to make a deep copy of the
packet before sending between hosts, to guarantee that the packet
is not shared. This introduces memory overhead since all payloads
will be duplicated.

The next commit will reduce memory overhead by only duplicating
the packet headers, but sharing the packet payload among multiple
versions of the same packet. This is more appropriate since the
networking layer will never modify the payload contents anyway.

refs #666