Don't copy packet payloads when sending across hosts
This commit decouples the payload data from a packet. Packets are
no longer shared among multiple hosts, and so all packet mutexes
are removed. Packet payloads can be shared and mutexes are added
to the payload structs to protect access.

This reduces memory overhead introduced in the previous commit
since payloads are no longer copied for every packet that gets
sent.

refs #666